Add Card Generation

+++ Src/Domain/Entities/Game/Items/Trinket.cs
@@ -23,7 +23,7 @@
 
         public string Effect { get; set; }
 
-        public int EffectPower { get; set; }
+        public double EffectPower { get; set; }
 
         public int Duration { get; set; }
